- api resources in a package for easier collaboration, - get rid of id_parser by putting single APIs in /<int:id> - use get(table, id) instead of filter_by
