+++ RecipeAPI/Resources/__init__.py
@@ -1,0 +1,4 @@
+from .units_resource import UnitsAPI, UnitAPI
+from .ingredients_resource import IngredientsAPI, IngredientAPI
+from .cuisines_resource import CuisinesAPI, CuisineAPI
+from .recipes_resource import RecipesAPI, RecipeAPI
