+++ RecipeAPI/Resources/ingredients_resource.py
@@ -1,0 +1,69 @@
+from flask import request
+from flask_restful import Resource, reqparse
+import jsons
+from Dtos import IngredientsDto, IngredientDto
+import database_service.api_helper as helper
+
+ingredients_parser = reqparse.RequestParser()
+ingredients_parser.add_argument('ingredient', type=str, action='append')
+ingredients_parser.add_argument('description', type=str, action='append')
+
+
+class IngredientsAPI(Resource):
+    def get(self):
+        """GET a collection of ingredients"""
+        allargs = ingredients_parser.parse_args()
+        args = {}
+        for key, value in allargs.items():
+            if value is not None:
+                args[key] = value
+        Ingredients = helper.GetIngredientsWithArguments(args)
+        return jsons.dump(Ingredients, sort_keys=False)
+
+    def post(self):
+        """POST a collection of ingredients or a single ingredient item"""
+        #try loading the collection
+        try:
+            Ingredients = jsons.load(request.get_json(force=True), cls=IngredientsDto, strict=True)
+            iscollection = True
+        except:
+            iscollection = False
+        #if not a collection load a single item, returns 404 when error
+        if not iscollection:
+            Ingredient = jsons.load(request.get_json(force=True), cls=IngredientDto, strict=True)
+        if iscollection:
+            if helper.AddIngredients(Ingredients):
+                return {"message": "created"}, 201
+            else:
+                return {"message": "conflict"}, 409
+        else:
+            if helper.AddIngredient(Ingredient):
+                return {"message": "created"}, 201
+            else:
+                return {"message": "conflict"}, 409
+
+
+class IngredientAPI(Resource):
+    def get(self, id):
+        """GET a single item by id"""
+        Ingredient = helper.GetIngredientById(id)
+        if Ingredient is not None:
+            return jsons.dump(Ingredient, sort_keys=False)
+        else:
+            return {"message": "not found"}, 404
+        
+    def put(self, id):
+        """PUT(update) a single item by id"""
+        Ingredient = jsons.load(request.get_json(force=True), cls=Dtos.IngredientDto, strict=True)
+        if helper.UpdateIngredient(Ingredient, id):
+            return {"message": "updated"}, 
+        else:
+            return {"message": "conflict"}, 409
+
+    def delete(self, id):
+        """DELETE a single item by id"""
+        if helper.DeleteIngredient(id):
+            return {"message": "deleted"}
+        else:
+            return {"message": "not allowed"}
+
